+++ app/src/main/java/com/example/reposearchapp/di/DataModule.kt
@@ -60,89 +60,3 @@ object GithubLanguageColorUtilModule {
             defaultDispatcher
         )
 }
-
-@Module
-@InstallIn(SingletonComponent::class)
-object ConverterModule {
-
-    @Provides
-    fun provideJsonConverter(
-    ): Converter.Factory {
-        return Json {
-            ignoreUnknownKeys = true
-        }.asConverterFactory(MediaType.parse("application/json")!!)
-    }
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-object InterceptorModule {
-
-    @Provides
-    fun provideInterceptor(
-        accessTokenRepository: AccessTokenRepository
-    ): GithubServiceInterceptor {
-        return GithubServiceInterceptor(accessTokenRepository)
-    }
-}
-
-@Qualifier
-annotation class AccessApiOkHttpClient
-
-@Qualifier
-annotation class GithubApiOkHttpClient
-
-@Module
-@InstallIn(SingletonComponent::class)
-object OkHttpModule {
-
-    @AccessApiOkHttpClient
-    @Provides
-    fun provideAccessApiOkHttpClient(
-    ): OkHttpClient {
-        return OkHttpClient().newBuilder().build()
-    }
-
-    @GithubApiOkHttpClient
-    @Provides
-    fun provideGithubApiOkHttpClient(
-        interceptor: GithubServiceInterceptor
-    ): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor(interceptor)
-            .build()
-    }
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    fun provideAccessApi(
-        @AccessApiOkHttpClient
-        okHttpClient: OkHttpClient,
-        converter: Converter.Factory
-    ): AccessApi {
-        return Retrofit.Builder()
-            .client(okHttpClient)
-            .addConverterFactory(converter)
-            .baseUrl("https://github.com/")
-            .build()
-            .create(AccessApi::class.java)
-    }
-
-    @Provides
-    fun provideGithubApi(
-        @GithubApiOkHttpClient
-        okHttpClient: OkHttpClient,
-        converter: Converter.Factory
-    ): GithubApi {
-        return Retrofit.Builder()
-            .client(okHttpClient)
-            .addConverterFactory(converter)
-            .baseUrl("https://api.github.com")
-            .build()
-            .create(GithubApi::class.java)
-    }
-}
